refactor: Rename `lib/plugins/aws/utils/getS3ObjectsFromStacks.js`
