+++ lib/plugins/aws/deploy/lib/cleanupS3Bucket.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const findAndGroupDeployments = require('../../utils/find-and-group-deployments');
-const getS3ObjectsFromStacks = require('../../utils/getS3ObjectsFromStacks');
+const getS3ObjectsFromStacks = require('../../utils/get-s3-objects-from-stacks');
 const { log } = require('@serverless/utils/log');
 
 module.exports = {
