+++ lib/plugins/aws/utils/get-s3-objects-from-stacks.js
